Move around more code for fewer dependencies?
