+++ core/include/scipp/core/except.h
@@ -14,6 +14,7 @@
 #include "scipp/common/except.h"
 #include "scipp/common/index.h"
 #include "scipp/core/dtype.h"
+#include "scipp/core/string.h"
 #include "scipp/units/except.h"
 #include "scipp/units/unit.h"
